[FIX] 검색기능 헤더에서 게시판으로 이동
